+++ include/jemalloc/internal/arena_inlines_a.h
@@ -2,7 +2,7 @@
 #define JEMALLOC_INTERNAL_ARENA_INLINES_A_H
 
 #include "jemalloc/internal/jemalloc_preamble.h"
-#include "jemalloc/internal/arena_structs.h"
+#include "jemalloc/internal/arena.h"
 
 static inline unsigned
 arena_ind_get(const arena_t *arena) {
