+++ include/jemalloc/internal/large.h
@@ -4,7 +4,8 @@
 #include "jemalloc/internal/jemalloc_preamble.h"
 #include "jemalloc/internal/edata.h"
 
-/* Forward decl; only prof_info_t * is used as a pointer arg below. */
+/* Forward decls; only used as pointer types below. */
+typedef struct arena_s     arena_t;
 typedef struct prof_info_s prof_info_t;
 
 void *large_malloc(tsdn_t *tsdn, arena_t *arena, size_t usize, bool zero);
